CRUD Search and User Login Logout and Registration

+++ studybud/base/views.py
@@ -1,7 +1,14 @@
 from django.shortcuts import render, redirect
 from django.http import HttpResponse
+from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth.decorators import login_required
 
-from .models import Room
+
+from .models import Room, Topic
 from .form import RoomForm
 
 
@@ -13,17 +20,72 @@ from .form import RoomForm
 #     {'id':3, 'name':'front end developers'} ,
 # ]
 
+def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist')
+
+        user = authenticate(request, username = username, password = password)
+        if user is not None:
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or password does not exist')
+
+    context = {'page':page}
+    return render(request,'base/login_register.html',context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerUser(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit = False)
+            user.username = user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,'An error occurred during registration')
+    return render(request, 'base/login_register.html', {'form':form})
 
 def home(request):
-    rooms = Room.objects.all()
-    context = {'rooms':rooms}
+    q = request.GET.get('q') if request.GET.get('q') != None else ''
+
+    rooms = Room.objects.filter(
+        Q(topic__name__icontains = q) |
+        Q(name__icontains = q) |
+        Q(description__icontains = q)
+        )
+
+
+    topics = Topic.objects.all()
+    room_count = rooms.count()
+    context = {'rooms':rooms,'topics':topics,'room_count':room_count}
     return render(request, 'base/home.html' , context)
 
 def room(request,pk):
     room = Room.objects.get(id = pk)
-    context = {'room':room}
+    room_messages = room.message_set.all().order_by('-created')
+    context = {'room':room,'room_messages':room_messages}
     return render(request, 'base/room.html', context)
 
+
+@login_required(login_url='login')
 def CreateRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -35,9 +97,16 @@ def CreateRoom(request):
     context = {'form':form}
     return render(request,'base/room_form.html',context)
 
+
+@login_required(login_url='login')
 def UpdateRoom(request,pk):
     room = Room.objects.get(id = pk)
     form = RoomForm(instance = room)
+
+    if request.user != room.host:
+        return HttpResponse('You do not have permission to edit this room')
+
+
     if request.method == 'POST':
         form = RoomForm(request.POST,instance=room)
         if form.is_valid():
@@ -47,8 +116,14 @@ def UpdateRoom(request,pk):
     context = {'form':form}
     return render(request,'base/room_form.html', context)
 
+
+@login_required(login_url='login')
 def DeleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('Your not allowed to delete this room')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
